refactor(#2): rename test/utils to test/util for consistency
